src/backend/task/requirements.txt

amazon_transcribe 
boto3
pytest-playwright 
requests
